feat: 매너 온도를 걷어내고 티어 점수를 정수 산식으로 계산한다
점수 및 티어 정책 §4 + 티어·점수 백엔드 테크 스펙. 36.5·EMA·밴드 체계와 티어 축이 병존하던
상태를 끝낸다 — 지금까지는 티어를 읽는 화면과 온도를 쓰는 경로가 따로 살아 있었다.

## 소수를 만들지 않는다

주간 배점 W 를 목표 횟수 N 으로 나누면 소수가 나온다(골드 주 7회 = 0.857…). 그 소수를 저장하는
대신 k번째 확정까지의 반영 누계를 정수로 직접 계산한다.

    f(k) = ⌊(2·W·k + N) ÷ (2·N)⌋

W×k÷N 의 사사오입과 동치이면서 나눗셈까지 전부 정수 연산이다. 언어 기본 round() 는 은행가
반올림일 수 있고 double 은 재현성이 없어 둘 다 쓰지 않는다. 이 식이 주는 성질이 셋이다.
멱등(카운트만의 함수라 두 번 처리해도 같다), 소급 정정이 쉬움(카운트를 고쳐 다시 계산),
그리고 f(N) = W 라서 마감 보정 절차 자체가 필요 없다. 5티어 × 2축 × 주 1~7회 전 조합을
단위 테스트로 고정했다.

## 한도는 챌린지별 각 사이클에 건다

달력 주차도 계정 합산도 아니다. 상태가 사이클 행 하나에 들어 있어 주차 키도 계정 단위 한도
테이블도 필요 없고, "어느 챌린지가 한도를 썼나"가 행 자체로 답이 된다. 상승과 하락을 따로 세지
않고 원점수 누계 하나를 클램핑하므로, 한도에 닿았다 되돌아오는 구간이 자연스럽게 처리된다 —
+25 → −5 면 순변동이 여전히 +20 이라 반영은 그대로고, +25 → −10 이면 +15 로 함께 내려온다.

핵심은 limitedCumulative 를 target 으로 덮어쓰지 않고 appliedDelta 만큼만 전진시키는 것이다.
덮어쓰면 0점 계정에서 반영되지도 않은 감점이 한도를 소비하고, 이후 원점수가 회복될 때
받은 적 없는 점수가 지급된다.

## 이벤트가 아니라 재정산

성공에는 별도 도메인 이벤트가 없고 확정 경로가 여러 곳이다. 그래서 이벤트를 하나하나 잡는 대신
판정 원본에서 카운트를 다시 세고 차이만 반영한다 — 정책 §4.5 가 명시적으로 허락하는 방식이다.
누락도 중복도 구조적으로 불가능해지고, 이의 인용 정정이 같은 함수를 다시 부르는 것으로 끝난다.
역분개 원장을 따로 짤 필요가 없다.

워터마크는 별도 테이블이 아니라 cycle_score_states.last_judged_at 의 최댓값이다 —
정산 대상 자체에 "어디까지 봤나"가 적혀 있어 상태가 두 벌로 갈라지지 않는다.

## 수동 인증은 두 축 어디에도 넣지 않는다

성공만 빼고 인증 가능일에서는 소모된 것으로 세면 만회할 날이 줄어 미달이 확정된다 —
수동 인증이 감점으로 둔갑한다. 그래서 그 날 자체를 없는 것처럼 건너뛴다.

## 사건성 감점

부정행위 −50 · 권한 미허용 강퇴 −15 · 중도 탈퇴 −⌈15 × (1 − 진행주간/52)⌉. 사이클 한도를
거치지 않고 즉시 전액 반영하며 연속 기록도 건드리지 않는다. 연속 실패 강퇴는 감점이 없다 —
각 주의 루틴 점수에 이미 반영돼 있어 중복 부과하지 않는다.

중도 탈퇴는 지금까지 −15 상수를 응답에만 실어 보내고 실제로 깎지 않았다. 이제 진행 기간에
반비례하는 실제 감점으로 반영한다.

## V32

cycle_score_states · challenge_streaks · score_corrections 신설, score_transactions 를
raw/limited/applied 3단 원장으로 재편. ReputationScore · ReputationSnapshot · Milestone 삭제,
challenges.min_manner_temperature 삭제 — 게이팅은 이미 표시 티어(min_tier)로 하고 있다.

+++ src/main/java/com/ruleup/ruleup_backend/me/service/MeTierHistoryService.java
@@ -74,7 +74,7 @@ public class MeTierHistoryService {
 
     /** 월말 스냅샷 — 그 달 마지막 변동의 잔액. 변동이 없던 달은 점이 없다(그래프가 직선으로 잇는다). */
     private List<MeTierHistoryResponse.Monthly> monthly(List<ScoreTransaction> ledger) {
-        Map<YearMonth, Long> lastOfMonth = new LinkedHashMap<>();
+        Map<YearMonth, Integer> lastOfMonth = new LinkedHashMap<>();
         for (ScoreTransaction t : ledger) {
             YearMonth ym = YearMonth.from(LocalDate.ofInstant(t.getCreatedAt(), KST));
             lastOfMonth.put(ym, t.getBalanceAfter());   // 오래된 순이라 마지막 put 이 월말 값이다
